refactor(src): split infra/config.rs by responsibility
