+++ src/infra/config/mod.rs
@@ -17,6 +17,8 @@
 //! Public field list, semantics, and defaults are pinned by
 //! `CONTRACTS.md` — every field documented there must be present.
 
+#![allow(unused_imports)]
+
 use std::collections::{HashMap, HashSet};
 use std::path::{Path, PathBuf};
 
@@ -1015,828 +1017,18 @@ impl Config {
 }
 
 // ---------------------------------------------------------------------------
-// Resolution chain
-// ---------------------------------------------------------------------------
 
-/// Where the configuration came from. Used in error messages so the
-/// operator can tell which level of the chain produced the failure.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ResolvedSource {
-    /// `$CADUCEUS_CONFIG` was set and pointed at this file.
-    ExplicitEnv,
-    /// `$HERMES_HOME/config.yaml` had a `caduceus:` section.
-    HermesHome,
-    /// `~/.config/caduceus/config.yaml` was the only one present.
-    Standalone,
-}
+// Submodule declarations and re-exports. The public surface keeps
+// `Config`, `RawConfig`, `OciPullPolicy`, etc. at `crate::infra::config::*`.
 
-/// Compute the list of files to consider, in order, given the three
-/// optional inputs from the loader.
-fn resolve_sources(
-    env: Option<&Path>,
-    hermes: Option<&Path>,
-    standalone: Option<&Path>,
-) -> CaduceusResult<Vec<(ResolvedSource, std::path::PathBuf)>> {
-    let mut sources: Vec<(ResolvedSource, std::path::PathBuf)> = Vec::new();
-    if let Some(path) = env {
-        let expanded = expand_leading_tilde(path.to_path_buf());
-        sources.push((ResolvedSource::ExplicitEnv, expanded));
-    }
-    if let Some(hermes_home) = hermes {
-        if hermes_home.as_os_str().is_empty() {
-            return Err(CaduceusError::Config(
-                "HERMES_HOME must not be empty".to_string(),
-            ));
-        }
-        // Reject relative HERMES_HOME per the contract.
-        if hermes_home.is_relative() {
-            return Err(CaduceusError::Config(
-                "HERMES_HOME must be an absolute path".to_string(),
-            ));
-        }
-        sources.push((ResolvedSource::HermesHome, hermes_home.join("config.yaml")));
-    }
-    if let Some(path) = standalone {
-        sources.push((ResolvedSource::Standalone, path.to_path_buf()));
-    }
-    Ok(sources)
-}
+pub mod load;
+pub mod setup;
+pub mod token;
 
-/// Read the raw configuration from the first successful candidate.
-/// Hermes files without a ``caduceus:`` section are skipped only when
-/// a standalone source is also available.
-fn load_raw_from_candidates(
-    sources: &[(ResolvedSource, std::path::PathBuf)],
-) -> CaduceusResult<RawConfig> {
-    if sources.is_empty() {
-        return Err(CaduceusError::Config(
-            "no configuration source provided".to_string(),
-        ));
-    }
+use self::load::*;
+use self::setup::*;
+use self::token::*;
 
-    // An explicit $CADUCEUS_CONFIG is an authoritative request — a
-    // missing file is a hard error. The operator either meant for
-    // that path to exist or set the variable by mistake.
-    if let Some((ResolvedSource::ExplicitEnv, path)) = sources.first() {
-        if !path.is_file() {
-            return Err(CaduceusError::Config(format!(
-                "$CADUCEUS_CONFIG points at {} but the file is missing",
-                path.display()
-            )));
-        }
-        return load_raw_from(path).map_err(|err| match err {
-            CaduceusError::Yaml(yaml_err) => {
-                CaduceusError::Config(format!("failed to parse {}: {yaml_err}", path.display()))
-            }
-            other => other,
-        });
-    }
-
-    let mut standalone_seen = false;
-    let mut hermes_seen_without_section = false;
-    let mut last_missing_standalone: Option<&std::path::Path> = None;
-    for (source, path) in sources {
-        match source {
-            ResolvedSource::HermesHome => {
-                if !path.is_file() {
-                    continue;
-                }
-                match load_raw_from(path) {
-                    Ok(raw) => return Ok(raw),
-                    Err(CaduceusError::Config(msg))
-                        if msg.contains("missing 'caduceus:' section")
-                            || msg.contains("has no 'caduceus:' section") =>
-                    {
-                        // Hermes file present but no caduceus section.
-                        // Per the task, fall through only if a
-                        // standalone source also exists.
-                        hermes_seen_without_section = true;
-                    }
-                    Err(other) => return Err(other),
-                }
-            }
-            ResolvedSource::Standalone => {
-                standalone_seen = true;
-                if !path.is_file() {
-                    last_missing_standalone = Some(path.as_path());
-                    continue;
-                }
-                // If a previous Hermes file was missing the section,
-                // we still want the standalone file to take over.
-                return load_raw_from(path);
-            }
-            ResolvedSource::ExplicitEnv => unreachable!(),
-        }
-    }
-
-    if hermes_seen_without_section && !standalone_seen {
-        return Err(CaduceusError::Config(
-            "Hermes config has no 'caduceus:' section and no standalone config was found"
-                .to_string(),
-        ));
-    }
-    if hermes_seen_without_section {
-        // Standalone was configured but the file was missing — surface
-        // that explicitly so the operator can fix it.
-        let path = last_missing_standalone
-            .map(|p| p.display().to_string())
-            .unwrap_or_else(|| "<unset>".to_string());
-        return Err(CaduceusError::Config(format!(
-            "Hermes config has no 'caduceus:' section and standalone config {path} is missing"
-        )));
-    }
-
-    Err(CaduceusError::Config(
-        "no configuration source found (set $CADUCEUS_CONFIG, $HERMES_HOME, or write ~/.config/caduceus/config.yaml)".to_string(),
-    ))
-}
-
-/// Read a single configuration file. Detects the Hermes shape (a top
-/// level ``caduceus:`` mapping) and unwraps it before deserialising
-/// into [`RawConfig`].
-fn load_raw_from(path: &Path) -> CaduceusResult<RawConfig> {
-    let text = std::fs::read_to_string(path).map_err(|err| {
-        CaduceusError::Config(format!("failed to read {}: {err}", path.display()))
-    })?;
-    parse_raw_from_text(&text, path)
-}
-
-fn parse_raw_from_text(text: &str, source_path: &Path) -> CaduceusResult<RawConfig> {
-    let outer: serde_yaml::Value = serde_yaml::from_str(text)?;
-    let map = outer.as_mapping().ok_or_else(|| {
-        CaduceusError::Config(format!(
-            "expected a YAML mapping at the root of {}",
-            source_path.display()
-        ))
-    })?;
-    if map.contains_key("caduceus") {
-        // Hermes-shaped file: extract the ``caduceus:`` mapping.
-        let section = map.get("caduceus").ok_or_else(|| {
-            CaduceusError::Config(format!(
-                "missing 'caduceus:' section in {}",
-                source_path.display()
-            ))
-        })?;
-        let raw: RawConfig = serde_yaml::from_value(section.clone())?;
-        return Ok(raw);
-    }
-    // Standalone-shaped file: every top-level key is part of the
-    // raw config. We rely on ``deny_unknown_fields`` to catch
-    // typos and stray sections — but only if the keys look like
-    // Caduceus config. Detect Hermes-style keys (which the contract
-    // expects on the same host) and treat the missing ``caduceus:``
-    // section as an explicit error rather than a parse failure.
-    for key in map.keys() {
-        if let Some(name) = key.as_str() {
-            if matches!(
-                name,
-                "model"
-                    | "agent"
-                    | "providers"
-                    | "tools"
-                    | "memory"
-                    | "cron"
-                    | "platforms"
-                    | "gateway"
-                    | "secrets"
-                    | "voice"
-                    | "mcp"
-                    | "tts"
-            ) {
-                return Err(CaduceusError::Config(format!(
-                    "Hermes config at {} has no 'caduceus:' section",
-                    source_path.display()
-                )));
-            }
-        }
-    }
-    let raw: RawConfig = serde_yaml::from_str(text)?;
-    Ok(raw)
-}
-
-/// Apply ``CADUCEUS_DRY_RUN`` to a parsed Config.
-fn apply_dry_run_env(cfg: &mut Config, value: &str) -> CaduceusResult<()> {
-    let normalized = value.trim().to_ascii_lowercase();
-    match normalized.as_str() {
-        "1" | "true" | "yes" => cfg.dry_run = true,
-        "0" | "false" | "no" => cfg.dry_run = false,
-        _ => {
-            return Err(CaduceusError::Config(format!(
-                "CADUCEUS_DRY_RUN must be one of 1/true/yes/0/false/no (got {value:?})"
-            )));
-        }
-    }
-    Ok(())
-}
-
-// ---------------------------------------------------------------------------
-// Token resolution
-// ---------------------------------------------------------------------------
-
-/// Indicate which resolution path produced the token. Used in tests
-/// and in the daemon's structured logs (without the secret itself).
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum TokenSource {
-    ExplicitConfig,
-    CaduceusEnv,
-    GithubEnv,
-    GhCli,
-}
-
-/// Environment variable lookup abstracted over the host process or a
-/// test fixture. Implementations must never leak the value through
-/// the trait surface — only non-secret metadata.
-pub trait TokenEnv {
-    /// Read an environment variable, returning `None` when unset or
-    /// empty. Whitespace-only values are also treated as unset.
-    fn get(&self, name: &str) -> Option<String>;
-}
-
-/// Process-environment adapter. Reads from the real OS env via
-/// `std::env::var_os`. Wrapped in a struct so tests can swap in a
-/// fake without mutating process state under concurrent tests.
-#[derive(Clone, Copy, Debug, Default)]
-pub struct OsEnv;
-
-impl TokenEnv for OsEnv {
-    fn get(&self, name: &str) -> Option<String> {
-        std::env::var_os(name)
-            .map(|value| value.to_string_lossy().trim().to_string())
-            .filter(|value| !value.is_empty())
-    }
-}
-
-/// Run a `gh auth token` subprocess with a 10-second timeout, captured
-/// stderr, and no token logging. The runner is overridable in tests.
-pub trait GhRunner: Send + Sync {
-    fn run(&self) -> Result<GhRunnerOutput, CaduceusError>;
-}
-
-/// What `gh auth token` produced, reduced to the contract surface.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct GhRunnerOutput {
-    pub exit_status: i32,
-    pub stdout: String,
-    pub stderr: String,
-}
-
-/// Default `gh` runner. Resolves the binary, shells out with an
-/// argument array, and surfaces exit codes / stderr without echoing
-/// the captured stdout.
-#[derive(Debug)]
-pub struct RealGhRunner;
-
-impl GhRunner for RealGhRunner {
-    fn run(&self) -> Result<GhRunnerOutput, CaduceusError> {
-        // ``which::which`` is the contract-respecting binary
-        // resolver; absent ``gh`` is a clean error.
-        let binary = match which::which("gh") {
-            Ok(path) => path,
-            Err(_) => {
-                return Err(CaduceusError::TokenResolution(
-                    "`gh` executable not found in PATH".to_string(),
-                ));
-            }
-        };
-        // ``subprocess::Command`` requires async + tokio; for the
-        // single-shot blocking 10-second call we use ``std::process``
-        // which is enough and avoids tying the resolver to a runtime.
-        // We do *not* log stdout — by contract the value is secret.
-        let mut command = std::process::Command::new(&binary);
-        command.arg("auth").arg("token");
-        command.env_clear();
-        // Inherit only PATH-equivalent vars the binary needs. We
-        // deliberately do not inherit the daemon's GitHub token so
-        // the operator's existing ``gh auth login`` state is the
-        // single source of truth. HOME is needed for ``gh`` to find
-        // its config directory.
-        for var in ["PATH", "HOME", "USER", "XDG_CONFIG_HOME"] {
-            if let Some(value) = std::env::var_os(var) {
-                command.env(var, value);
-            }
-        }
-        let output = match command.output() {
-            Ok(out) => out,
-            Err(err) => {
-                return Err(CaduceusError::TokenResolution(format!(
-                    "failed to spawn `gh`: {err}"
-                )));
-            }
-        };
-        let exit_status = output.status.code().unwrap_or(-1);
-        let stdout = String::from_utf8_lossy(&output.stdout).to_string();
-        let stderr = String::from_utf8_lossy(&output.stderr).to_string();
-        Ok(GhRunnerOutput {
-            exit_status,
-            stdout,
-            stderr,
-        })
-    }
-}
-
-/// Resolved token + which source produced it.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ResolvedToken {
-    pub token: String,
-    pub source: TokenSource,
-}
-
-impl ResolvedToken {
-    /// Bundle a token with its source for callers that want to log
-    /// the resolution path without exposing the secret.
-    pub fn new(token: String, source: TokenSource) -> Self {
-        Self { token, source }
-    }
-}
-
-/// Implementation of the documented hierarchy. Public so tests can
-/// drive it with their own env / gh fixtures.
-pub fn resolve_token_chain(
-    cfg: &Config,
-    env: &dyn TokenEnv,
-    runner: &dyn GhRunner,
-) -> CaduceusResult<ResolvedToken> {
-    if let Some(token) = non_empty(cfg.github_token.as_deref()) {
-        return Ok(ResolvedToken::new(token, TokenSource::ExplicitConfig));
-    }
-    if let Some(token) = env.get("CADUCEUS_GITHUB_TOKEN") {
-        return Ok(ResolvedToken::new(token, TokenSource::CaduceusEnv));
-    }
-    if let Some(token) = env.get("GITHUB_TOKEN") {
-        return Ok(ResolvedToken::new(token, TokenSource::GithubEnv));
-    }
-
-    // Final fallback: ``gh auth token``.
-    match runner.run() {
-        Ok(out) if out.exit_status == 0 => {
-            let trimmed = out.stdout.trim().to_string();
-            if is_token_usable(&trimmed) {
-                return Ok(ResolvedToken::new(trimmed, TokenSource::GhCli));
-            }
-            Err(CaduceusError::TokenResolution(
-                "`gh auth token` returned no usable token".to_string(),
-            ))
-        }
-        Ok(out) => Err(CaduceusError::TokenResolution(format!(
-            "`gh auth token` exited {} (stderr suppressed)",
-            out.exit_status
-        ))),
-        Err(err) => Err(err),
-    }
-}
-
-/// Return ``Some(token)`` when *token* is non-empty after trimming and
-/// contains at least one non-whitespace character.
-fn is_token_usable(token: &str) -> bool {
-    !token.trim().is_empty()
-}
-
-fn non_empty(s: Option<&str>) -> Option<String> {
-    s.map(str::trim)
-        .filter(|t| !t.is_empty())
-        .map(str::to_string)
-}
-
-// ---------------------------------------------------------------------------
-// Path / interpolation helpers
-// ---------------------------------------------------------------------------
-
-/// Expand a single leading ``~`` to the user's home directory. No
-/// other shell expansion is performed (CONTRACTS.md "Configuration").
-pub fn expand_leading_tilde(path: PathBuf) -> PathBuf {
-    if let Ok(rest) = path.strip_prefix("~") {
-        if let Some(home) = home_dir() {
-            if rest.as_os_str().is_empty() {
-                return home;
-            }
-            // Preserve a literal separator so ``~/foo`` does not
-            // collapse to ``~foo`` when ``home`` already ends with a
-            // separator.
-            let mut combined = home;
-            if !rest.starts_with("/") && !rest.starts_with("\\") {
-                combined.push(std::path::Path::new(rest));
-            } else {
-                combined.push(rest);
-            }
-            return combined;
-        }
-    }
-    path
-}
-
-fn home_dir() -> Option<PathBuf> {
-    if let Some(raw) = std::env::var_os("HOME") {
-        if !raw.is_empty() {
-            return Some(PathBuf::from(raw));
-        }
-    }
-    if let Some(profile) = std::env::var_os("USERPROFILE") {
-        if !profile.is_empty() {
-            return Some(PathBuf::from(profile));
-        }
-    }
-    None
-}
-
-/// Try to discover the Hermes plugin root directory.
-///
-/// When Caduceus is installed as a Hermes plugin, the plugin directory
-/// sits at `$HERMES_HOME/plugins/caduceus/`. Returns `Some(...)` only
-/// when that directory exists and contains the canonical plugin-assets
-/// layout, so a stale empty directory from a partial install does not
-/// trigger plugin-root resolution.
-fn discover_plugin_root(hermes_home: &Path) -> Option<PathBuf> {
-    let candidate = hermes_home.join("plugins").join("caduceus");
-    if !candidate.is_dir() {
-        return None;
-    }
-    // Validate the install by checking for the plugin-assets bridge.
-    if candidate
-        .join("plugin-assets")
-        .join("worker-bridge.py")
-        .is_file()
-    {
-        Some(candidate)
-    } else {
-        None
-    }
-}
-
-fn default_state_dir(ctx: &LoadContext) -> PathBuf {
-    if let Some(ref h) = ctx.hermes_home {
-        return h.join("caduceus-state");
-    }
-    expand_leading_tilde(PathBuf::from("~/.hermes/caduceus-state"))
-}
-
-fn default_workdir_base(ctx: &LoadContext) -> PathBuf {
-    // Hermes-primary default lives under HERMES_HOME; standalone
-    // installs typically use ~/projects. We default to the latter
-    // unless HERMES_HOME is set, in which case the Hermes-managed
-    // <HERMES_HOME>/projects path is used.
-    if let Some(ref h) = ctx.hermes_home {
-        return h.join("projects");
-    }
-    expand_leading_tilde(PathBuf::from("~/projects"))
-}
-
-fn default_worker_command(ctx: &LoadContext) -> Option<Vec<String>> {
-    // Priority 1: Hermes plugin layout — template bridge under plugin-assets/.
-    if let Some(ref plugin_root) = ctx.plugin_root {
-        let bridge = plugin_root.join("plugin-assets").join("worker-bridge.py");
-        return Some(vec![
-            "python3".to_string(),
-            bridge.to_string_lossy().to_string(),
-        ]);
-    }
-    // Priority 2: Hermes-primary install — user-owned bridge (AC-04).
-    if let Some(ref hermes_home) = ctx.hermes_home {
-        let bridge = hermes_home.join("caduceus").join("worker-bridge.py");
-        return Some(vec![
-            "python3".to_string(),
-            bridge.to_string_lossy().to_string(),
-        ]);
-    }
-    None
-}
-
-fn expand_worker_command(cmd: Vec<String>, ctx: &LoadContext) -> CaduceusResult<Vec<String>> {
-    let mut out = Vec::with_capacity(cmd.len());
-    for (idx, arg) in cmd.into_iter().enumerate() {
-        if arg.contains(PLUGIN_ROOT_TOKEN) {
-            if idx == 0 {
-                return Err(CaduceusError::Config(
-                    "${plugin_root} cannot appear in the program position of worker_command"
-                        .to_string(),
-                ));
-            }
-            let plugin_root = ctx.plugin_root.as_ref().ok_or_else(|| {
-                CaduceusError::Config(
-                    "${plugin_root} referenced but plugin root is not known".to_string(),
-                )
-            })?;
-            let replaced = arg.replace(PLUGIN_ROOT_TOKEN, &plugin_root.to_string_lossy());
-            out.push(replaced);
-        } else if arg.contains('$') || arg.contains('~') {
-            return Err(CaduceusError::Config(format!(
-                "worker_command argument {idx} contains unsupported interpolation: {arg:?}"
-            )));
-        } else {
-            out.push(arg);
-        }
-    }
-    Ok(out)
-}
-
-// ---------------------------------------------------------------------------
-// Validators
-// ---------------------------------------------------------------------------
-
-/// Validate a secret grant name: must match `[a-z][a-z0-9-]{0,63}`.
-fn is_valid_secret_grant_name(name: &str) -> bool {
-    if name.is_empty() {
-        return false;
-    }
-    let bytes = name.as_bytes();
-    if !bytes[0].is_ascii_lowercase() {
-        return false;
-    }
-    if name.len() > 64 {
-        return false;
-    }
-    bytes[1..]
-        .iter()
-        .all(|b| b.is_ascii_lowercase() || b.is_ascii_digit() || *b == b'-')
-}
-
-fn validate_watched_repos(repos: &[String], errors: &mut Vec<String>) {
-    let mut seen: HashSet<String> = HashSet::new();
-    for repo in repos {
-        let lower = repo.to_ascii_lowercase();
-        if seen.contains(&lower) {
-            errors.push(format!(
-                "watched_repos contains case-insensitive duplicate: {repo}"
-            ));
-            continue;
-        }
-        seen.insert(lower);
-        if !is_valid_repo_slug(repo) {
-            errors.push(format!("watched_repos entry is not owner/repo: {repo:?}"));
-        }
-    }
-}
-
-pub fn is_valid_repo_slug(repo: &str) -> bool {
-    // Owner/repo: owner uses GitHub's alphanumeric/hyphen rules (1..=39
-    // chars, no leading/trailing hyphen); repo is 1..=100 chars from
-    // [A-Za-z0-9_.-] excluding "." and "..".
-    if repo.is_empty() {
-        return false;
-    }
-    let Some((owner, repo_name)) = repo.split_once('/') else {
-        return false;
-    };
-    if repo_name.contains('/') {
-        return false;
-    }
-    crate::github::issue::validate_owner(owner).is_ok()
-        && crate::github::issue::validate_repo(repo_name).is_ok()
-}
-
-/// Positive allowlist validator for the `api_base` configuration value.
-///
-/// Per `CONTRACTS.md` GH-001, `api_base` MUST be either the literal
-/// `https://api.github.com` or an `https://` URL whose path is
-/// `/api/v3` (or `/api/v3/...`). Anything else — `http://`,
-/// `bitbucket.example.com`, custom path-prefixed proxies, malformed
-/// URLs — is rejected with a configuration error.
-///
-/// Rules (evaluated in order):
-/// 1. Trim the input. Empty after trim → `Err("api_base must not be empty")`.
-/// 2. Parse as `url::Url`. Parse failure → `Err("api_base is not a valid URL")`.
-/// 3. Scheme MUST be `https`. Otherwise → `Err("api_base scheme must be https")`.
-/// 4. Host MUST be present and non-empty.
-/// 5. For `api.github.com` → path must be `/` or empty (SaaS).
-///    For any other host → path must be `/api/v3` or start with `/api/v3/` (GHES).
-/// 6. If all rules pass → `Ok(())`.
-///
-/// # Loopback accommodation (test-only)
-///
-/// The validator additionally accepts `http://` URLs whose host is a
-/// loopback address (`localhost`, `127.0.0.1`, or any address in
-/// `127.0.0.0/8`). This is a **test-only** accommodation: the
-/// integration test suite (`tests/integration/integration_test.rs`) spawns the
-/// real `caduceus` binary against a wiremock HTTP server, and
-/// configuring wiremock with a self-signed TLS cert is out of scope
-/// for this task. Loopback addresses are not routable, so the
-/// production security posture is unaffected — a production config
-/// pointing `api_base` at a loopback address would fail to reach
-/// GitHub, but it would not expose credentials or accept a
-/// non-GitHub endpoint.
-///
-/// Error messages NEVER include the raw `value` parameter to prevent
-/// credential leaks. They reference only parsed components or generic
-/// descriptions.
-///
-/// This function is a pure positive allowlist — it does NOT consult
-/// `comment_forbidden_strings` or any other forbidden-list to detect
-/// non-GitHub endpoints. The independence test in
-/// `tests/state/api_base_allowlist_test.rs` proves this property.
-pub fn validate_api_base(value: &str) -> Result<(), String> {
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return Err("api_base must not be empty".to_string());
-    }
-    let url = url::Url::parse(trimmed).map_err(|_| "api_base is not a valid URL".to_string())?;
-    let host = url.host_str().unwrap_or("");
-    let is_loopback = match url.host_str() {
-        Some("localhost") => true,
-        Some(h) => h == "127.0.0.1" || h.starts_with("127."),
-        _ => false,
-    };
-    if url.scheme() != "https" && !is_loopback {
-        return Err("api_base scheme must be https".to_string());
-    }
-    match url.host_str() {
-        Some(h) if !h.is_empty() => {}
-        _ => return Err("api_base must have a host".to_string()),
-    }
-    // Loopback: skip path validation (test-only).
-    if is_loopback {
-        return Ok(());
-    }
-    let path = url.path();
-    if host == "api.github.com" {
-        // GitHub.com SaaS: path must be empty or /
-        if path != "/" && !path.is_empty() {
-            return Err("api_base path must be / for api.github.com".to_string());
-        }
-    } else {
-        // GHES: path must be /api/v3 or /api/v3/...
-        if path != "/api/v3" && !path.starts_with("/api/v3/") {
-            return Err("api_base path must be /api/v3 for non-SaaS hosts".to_string());
-        }
-    }
-    Ok(())
-}
-
-fn validate_worker_command(cmd: &[String], errors: &mut Vec<String>) {
-    if cmd.is_empty() {
-        errors.push("worker_command must contain at least one argument".to_string());
-    }
-    for arg in cmd {
-        for forbidden in FORBIDDEN_INTERPOLATION_TOKENS {
-            if arg.contains(forbidden) {
-                errors.push(format!(
-                    "worker_command argument {arg:?} contains forbidden interpolation token {forbidden:?}"
-                ));
-            }
-        }
-    }
-}
-
-fn validate_comment_forbidden_strings(values: &[String], errors: &mut Vec<String>) {
-    for value in values {
-        if value.is_empty() {
-            errors.push("comment_forbidden_strings must not contain empty entries".to_string());
-        }
-    }
-}
-
-fn compile_ignore_patterns(
-    patterns: &[String],
-    errors: &mut Vec<String>,
-) -> CaduceusResult<Vec<Regex>> {
-    let mut compiled = Vec::with_capacity(patterns.len());
-    for pattern in patterns {
-        match Regex::new(pattern) {
-            Ok(re) => compiled.push(re),
-            Err(e) => errors.push(format!(
-                "comment_ignore_patterns contains invalid regex {pattern:?}: {e}"
-            )),
-        }
-    }
-    Ok(compiled)
-}
-
-fn validate_worker_env_allowlist(values: &[String], errors: &mut Vec<String>) {
-    for value in values {
-        validate_env_var_pattern(value, errors);
-        // Reject any pattern whose expansion could expose a denied
-        // credential name. Exact-name entries (``GITHUB_TOKEN``) and
-        // direct-prefix wildcards (``GITHUB_TOKEN_*``) are denied.
-        // Broader wildcards that include a denied name as a match
-        // (``GITHUB_*``) are also denied, because they would let
-        // ``GITHUB_TOKEN`` reach the worker through the wildcard.
-        if exposes_denied_credential(value) {
-            errors.push(format!(
-                "worker_env_allowlist contains denied credential name: {value:?}"
-            ));
-        }
-    }
-}
-
-/// Return ``true`` when *value* (an exact name or a terminal ``*``
-/// prefix wildcard) would expose any entry in [`DENIED_ENV_VARS`] to
-/// the worker process.
-pub(crate) fn exposes_denied_credential(value: &str) -> bool {
-    for denied in DENIED_ENV_VARS {
-        if value == *denied {
-            return true;
-        }
-        // The exact prefix wildcard (``NAME_*``) always exposes ``NAME``.
-        let prefix = format!("{denied}_*");
-        if value == prefix {
-            return true;
-        }
-        // A broader wildcard (``PREFIX_*``) exposes ``denied`` when
-        // ``denied`` starts with ``PREFIX`` (case-sensitive — the
-        // operator must opt in explicitly).
-        if let Some(prefix) = value.strip_suffix('*') {
-            if denied.starts_with(prefix) && denied.len() > prefix.len() {
-                return true;
-            }
-        }
-    }
-    false
-}
-
-/// A single env-var entry must be either an exact portable name or
-/// an exact terminal-`*` prefix pattern. Any other wildcard placement,
-/// empty entry, `=`, NUL, or non-portable character is rejected.
-pub(crate) fn validate_env_var_pattern(value: &str, errors: &mut Vec<String>) {
-    if value.is_empty() {
-        errors.push("worker_env_allowlist entry must not be empty".to_string());
-        return;
-    }
-    if value.contains('=') {
-        errors.push(format!(
-            "worker_env_allowlist entry {value:?} must not contain '='"
-        ));
-        return;
-    }
-    if value.contains('\0') {
-        errors.push(format!(
-            "worker_env_allowlist entry {value:?} must not contain NUL"
-        ));
-        return;
-    }
-    if !value
-        .chars()
-        .all(|c| c.is_ascii_alphanumeric() || c == '_' || c == '*')
-    {
-        errors.push(format!(
-            "worker_env_allowlist entry {value:?} contains non-portable characters"
-        ));
-        return;
-    }
-    // Only one terminal ``*`` is permitted.
-    if value.contains('*') {
-        let star_count = value.matches('*').count();
-        if star_count > 1 || !value.ends_with('*') {
-            errors.push(format!(
-                "worker_env_allowlist entry {value:?} may only contain a single terminal '*' wildcard"
-            ));
-        }
-    }
-    if exposes_denied_credential(value) {
-        errors.push(format!(
-            "worker_env_allowlist contains denied credential name: {value:?}"
-        ));
-    }
-}
-
-/// Validate the secure-path semantics for state-style directories:
-/// reject symlinks and refuse to accept paths that already exist as
-/// non-directories. The function does NOT touch the filesystem beyond
-/// `metadata`; creation is the daemon's job.
-fn validate_secure_path(path: &Path, field: &str, errors: &mut Vec<String>) {
-    if path.as_os_str().is_empty() {
-        errors.push(format!("{field} must not be empty"));
-        return;
-    }
-    if let Ok(meta) = std::fs::symlink_metadata(path) {
-        if meta.file_type().is_symlink() {
-            errors.push(format!("{field} must not be a symlink: {}", path.display()));
-        } else if !meta.is_dir() {
-            errors.push(format!(
-                "{field} exists but is not a directory: {}",
-                path.display()
-            ));
-        }
-    }
-}
-
-/// Validate `repo_storage_root`: refuse symlinks, and when the
-/// directory already exists refuse modes wider than 0700 with
-/// an operator-actionable fix message. Missing directories are
-/// accepted — the daemon creates them at startup with mode 0700.
-fn validate_repo_storage_root(path: &Path, errors: &mut Vec<String>) {
-    if path.as_os_str().is_empty() {
-        errors.push("repo_storage_root must not be empty".to_string());
-        return;
-    }
-    if let Ok(meta) = std::fs::symlink_metadata(path) {
-        if meta.file_type().is_symlink() {
-            errors.push(format!(
-                "repo_storage_root must not be a symlink: {}",
-                path.display()
-            ));
-            return;
-        }
-    }
-    if path.exists() {
-        if let Ok(meta) = std::fs::metadata(path) {
-            use std::os::unix::fs::PermissionsExt;
-            let mode = meta.permissions().mode() & 0o777;
-            if mode != 0o700 {
-                errors.push(format!(
-                    "repo_storage_root {} has mode {:03o}; expected 0700. Run: chmod 0700 {}",
-                    path.display(),
-                    mode,
-                    path.display()
-                ));
-            }
-        }
-    }
-}
+pub use load::*;
+pub use setup::*;
+pub use token::*;
